+++ examples/gpt_neo/deepspeed_hf.py
@@ -15,7 +15,7 @@ from slapo.logger import get_logger
 from slapo.op import ParallelCrossEntropy
 from slapo.utils.report import report_memory
 
-from slapo.model_schedule.gpt import schedule_model
+from slapo.model_schedule.gpt_neo import schedule_model
 from examples.utils import (
     train_with_deepspeed_engine,
     get_ds_config,
