+++ examples/gpt_neo/megatron_hf.py
@@ -44,7 +44,7 @@ def get_model(
     if "slapo" in impl:
         import slapo
         from slapo.utils.report import report_memory
-        from slapo.model_schedule.gpt import schedule_model
+        from slapo.model_schedule.gpt_neo import schedule_model
 
         report_memory()
         with slapo.init_empty_weights(enable=delay_init):
